Add testing functionality
Added jest module, allowing unit tests to be run. Currently, the code is very untestable. By nature, this code is difficult to test because it runs through the discord API. This does at least test some basic functions and will test that the cozy ids are being used instead of testing ids.